+++ config/config_test.go
@@ -60,6 +60,10 @@ func TestLoadConfig(t *testing.T) {
 		t.Error("should be true (config value should be used)")
 	}
 
+	if config.Filesystems.UseMountpoint != false {
+		t.Error("should be false (default value should be used)")
+	}
+
 	if config.Connection.PostMetricsDequeueDelaySeconds != 30 {
 		t.Error("should be 30 (default value should be used)")
 	}
@@ -108,6 +112,31 @@ func TestLoadConfigWithHostStatus(t *testing.T) {
 
 	if config.HostStatus.OnStop != "poweroff" {
 		t.Error(`HostStatus.OnStop should be "poweroff"`)
+	}
+}
+
+var sampleConfigWithMountPoint = `
+apikey = "abcde"
+display_name = "fghij"
+
+[filesystems]
+use_mountpoint = true
+`
+
+func TestLoadConfigWithMountPoint(t *testing.T) {
+	tmpFile, err := newTempFileWithContent(sampleConfigWithMountPoint)
+	if err != nil {
+		t.Errorf("should not raise error: %v", err)
+	}
+	defer os.Remove(tmpFile.Name())
+
+	config, err := LoadConfig(tmpFile.Name())
+	if err != nil {
+		t.Errorf("should not raise error: %v", err)
+	}
+
+	if config.Filesystems.UseMountpoint != true {
+		t.Error("should be true (config value should be used)")
 	}
 }
 
@@ -227,7 +256,7 @@ include = "%s/*.conf"
 command = "foo1"
 
 [plugin.metrics.bar]
-command = "this wille be overwritten"
+command = "this will be overwritten"
 `, tomlQuotedReplacer.Replace(configDir))
 
 	configFile, err := newTempFileWithContent(configContent)