+++ main.go
@@ -45,6 +45,13 @@ func main() {
 	if os.Getenv("GOMAXPROCS") == "" {
 		runtime.GOMAXPROCS(1)
 	}
+	// force disabling http2 for now
+	godebug := os.Getenv("GODEBUG")
+	if godebug != "" {
+		godebug += ","
+	}
+	godebug += "http2client=0"
+	os.Setenv("GODEBUG", godebug)
 	cli.Run(os.Args[1:])
 }
 
@@ -139,6 +146,10 @@ func resolveConfig(fs *flag.FlagSet, argv []string) (*config.Config, error) {
 
 	if conf.Apikey == "" {
 		return nil, fmt.Errorf("Apikey must be specified in the config file (or by the DEPRECATED command-line flag)")
+	}
+
+	if conf.HTTPProxy != "" {
+		os.Setenv("HTTP_PROXY", conf.HTTPProxy)
 	}
 	return conf, nil
 }