+++ metrics/linux/disk_test.go
@@ -17,7 +17,7 @@ func TestDiskGenerator(t *testing.T) {
 		t.Skipf("Skip: can't access `/proc/diskstats` in Travis environment.")
 	}
 
-	g := &DiskGenerator{1 * time.Second}
+	g := &DiskGenerator{Interval: 1 * time.Second}
 	values, err := g.Generate()
 	if err != nil {
 		t.Errorf("should not raise error: %v", err)
@@ -52,7 +52,8 @@ func TestParseDiskStats(t *testing.T) {
 
 202       2 xvda2 1641 9310 87552 1252 6365 3717 80664 24192 0 15040 25428`)
 
-	result, err := parseDiskStats(out)
+	var emptyMapping map[string]string
+	result, err := parseDiskStats(out, emptyMapping)
 	if err != nil {
 		t.Errorf("error should be nil but: %s", err)
 	}
@@ -83,5 +84,42 @@ func TestParseDiskStats(t *testing.T) {
 	}
 	if !reflect.DeepEqual(result, expect) {
 		t.Errorf("result is not expected one: %+v", result)
+	}
+
+	mapping := map[string]string{
+		"xvda1": "_some_mount",
+		"xvda3": "_nonused_mount",
+	}
+	resultWithMapping, err := parseDiskStats(out, mapping)
+	if err != nil {
+		t.Errorf("error should be nil but: %s", err)
+	}
+
+	expectWithMapping := metrics.Values{
+		"disk._some_mount.reads":          750193,
+		"disk._some_mount.readsMerged":    3037,
+		"disk._some_mount.sectorsRead":    28116978,
+		"disk._some_mount.readTime":       368712,
+		"disk._some_mount.writes":         16600606,
+		"disk._some_mount.writesMerged":   7233846,
+		"disk._some_mount.sectorsWritten": 424712632,
+		"disk._some_mount.writeTime":      23987908,
+		"disk._some_mount.ioInProgress":   0,
+		"disk._some_mount.ioTime":         2355636,
+		"disk._some_mount.ioTimeWeighted": 24345740,
+		"disk.xvda2.reads":                1641,
+		"disk.xvda2.readsMerged":          9310,
+		"disk.xvda2.sectorsRead":          87552,
+		"disk.xvda2.readTime":             1252,
+		"disk.xvda2.writes":               6365,
+		"disk.xvda2.writesMerged":         3717,
+		"disk.xvda2.sectorsWritten":       80664,
+		"disk.xvda2.writeTime":            24192,
+		"disk.xvda2.ioInProgress":         0,
+		"disk.xvda2.ioTime":               15040,
+		"disk.xvda2.ioTimeWeighted":       25428,
+	}
+	if !reflect.DeepEqual(resultWithMapping, expectWithMapping) {
+		t.Errorf("result is not expected one: %+v", resultWithMapping)
 	}
 }