+++ metrics/linux/memory.go
@@ -20,7 +20,7 @@ MemoryGenerator collect memory usage
 
 metric = "total", "free", "buffers", "cached", "active", "inactive", "swap_cached", "swap_total", "swap_free"
 
-Metrics "used" is caluculated here like (total - free - buffers - cached) for ease.
+Metrics "used" is calculated here like (total - free - buffers - cached) for ease.
 This calculation may be going to be done in server side in the future.
 
 graph: stacks `memory.{metric}`