+++ metrics/windows/filesystem.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/mackerelio/mackerel-agent/logging"
 	"github.com/mackerelio/mackerel-agent/metrics"
+	"github.com/mackerelio/mackerel-agent/util"
 	"github.com/mackerelio/mackerel-agent/util/windows"
 )
 
@@ -31,7 +32,7 @@ func (g *FilesystemGenerator) Generate() (metrics.Values, error) {
 	ret := make(map[string]float64)
 	for name, values := range filesystems {
 		if matches := regexp.MustCompile(`^(.*):`).FindStringSubmatch(name); matches != nil {
-			device := regexp.MustCompile(`[^A-Za-z0-9_-]`).ReplaceAllString(matches[1], "_")
+			device := util.SanitizeMetricKey(matches[1])
 
 			ret["filesystem."+device+".size"] = values.KbSize * 1024
 			ret["filesystem."+device+".used"] = values.KbUsed * 1024