+++ spec/cloud.go
@@ -13,10 +13,9 @@ import (
 )
 
 // This Generator collects metadata about cloud instances.
-// Currently only EC2 is supported.
+// Currently EC2 and GCE are supported.
 // EC2: http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/AESDG-chapter-instancedata.html
 // GCE: https://developers.google.com/compute/docs/metadata
-// DigitalOcean: https://developers.digitalocean.com/metadata/
 
 // CloudGenerator definition
 type CloudGenerator struct {
@@ -36,12 +35,11 @@ func (g *CloudGenerator) Key() string {
 
 var cloudLogger = logging.GetLogger("spec.cloud")
 
-var ec2BaseURL, gceMetaURL, digitalOceanBaseURL *url.URL
+var ec2BaseURL, gceMetaURL *url.URL
 
 func init() {
 	ec2BaseURL, _ = url.Parse("http://169.254.169.254/latest/meta-data")
 	gceMetaURL, _ = url.Parse("http://metadata.google.internal/computeMetadata/v1/?recursive=true")
-	digitalOceanBaseURL, _ = url.Parse("http://169.254.169.254/metadata/v1") // has not been yet used
 }
 
 var timeout = 100 * time.Millisecond
@@ -58,10 +56,18 @@ func SuggestCloudGenerator() *CloudGenerator {
 	return nil
 }
 
-func isEC2() bool {
-	cl := http.Client{
+func httpCli() *http.Client {
+	return &http.Client{
 		Timeout: timeout,
+		Transport: &http.Transport{
+			// don't use HTTP_PROXY when requesting cloud instance metadata APIs
+			Proxy: nil,
+		},
 	}
+}
+
+func isEC2() bool {
+	cl := httpCli()
 	// '/ami-id` is may be aws specific URL
 	resp, err := cl.Get(ec2BaseURL.String() + "/ami-id")
 	if err != nil {
@@ -78,9 +84,7 @@ func isGCE() bool {
 }
 
 func requestGCEMeta() ([]byte, error) {
-	cl := http.Client{
-		Timeout: timeout,
-	}
+	cl := httpCli()
 	req, err := http.NewRequest("GET", gceMetaURL.String(), nil)
 	if err != nil {
 		return nil, err
@@ -106,9 +110,7 @@ type EC2Generator struct {
 
 // Generate collects metadata from cloud platform.
 func (g *EC2Generator) Generate() (interface{}, error) {
-	client := http.Client{
-		Timeout: timeout,
-	}
+	cl := httpCli()
 
 	metadataKeys := []string{
 		"instance-id",
@@ -128,7 +130,7 @@ func (g *EC2Generator) Generate() (interface{}, error) {
 	metadata := make(map[string]string)
 
 	for _, key := range metadataKeys {
-		resp, err := client.Get(g.baseURL.String() + "/" + key)
+		resp, err := cl.Get(g.baseURL.String() + "/" + key)
 		if err != nil {
 			cloudLogger.Debugf("This host may not be running on EC2. Error while reading '%s'", key)
 			return nil, nil
@@ -156,9 +158,9 @@ func (g *EC2Generator) Generate() (interface{}, error) {
 
 // SuggestCustomIdentifier suggests the identifier of the EC2 instance
 func (g *EC2Generator) SuggestCustomIdentifier() (string, error) {
-	client := http.Client{Timeout: timeout}
+	cl := httpCli()
 	key := "instance-id"
-	resp, err := client.Get(g.baseURL.String() + "/" + key)
+	resp, err := cl.Get(g.baseURL.String() + "/" + key)
 	if err != nil {
 		return "", fmt.Errorf("Error while retrieving instance-id.")
 	}