+++ util/sanitize_test.go
@@ -1,0 +1,12 @@
+package util
+
+import "testing"
+
+func TestSanitizeMetricKey(t *testing.T) {
+	input := "abc*defあ.ggg"
+	expect := "abc_def__ggg"
+	output := SanitizeMetricKey(input)
+	if output != expect {
+		t.Errorf("invalid output of `SanitizeMetricKey`. expected: %s, output: %s", expect, output)
+	}
+}