+++ CHANGELOG.md
@@ -1,5 +1,31 @@
 # Changelog
 
+## 0.36.0 (2016-10-18)
+
+* don't use HTTP_PROXY when requesting cloud instance metadata APIs #285 (Songmu)
+* Add an option to output filesystem-related metrics with key by mountpoint #286 (astj)
+
+
+## 0.35.1 (2016-09-29)
+
+* support MACKEREL_PLUGIN_WORKDIR in init scripts #277 (Songmu)
+* Add platform metadata for Darwin #280 (astj)
+* Disable http2 for now #283 (Songmu)
+
+
+## 0.35.0 (2016-09-07)
+
+* built with Go 1.7 #266 (Songmu)
+* remove `func (vs *Values) Merge(other Values)` #268 (Songmu)
+* [incompatible] consider df  (used + available) as size of filesystem #271 (Songmu)
+* Remove DigitalOcean related comment/definition from spec/cloud.go #272 (astj)
+* Fix golint is not working on ci, and add some comment to pass golint #273 (astj)
+* Add linux distribution information to kernel spec #274 (ak1t0)
+* http_proxy configuration #275 (Songmu)
+* set PATH and LANG only in unix environment #276 (Songmu)
+* Ignore docker mapper storage in spec as well #278 (itchyny)
+
+
 ## 0.34.0 (2016-08-18)
 
 * Reduce retry count on finding a host by the custom identifier #258 (itchyny)